% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_mds}
\alias{plot_gene_mds}
\title{MDS plot of gene expression data}
\usage{
plot_gene_mds(
  expr_matrix,
  sample_labels = NULL,
  title = "MDS of Gene Expression",
  log_transform = FALSE,
  distance_method = "euclidean"
)
}
\arguments{
\item{expr_matrix}{Gene expression matrix (genes as rows, samples as columns)}

\item{sample_labels}{Optional vector of sample labels}

\item{title}{Plot title}

\item{log_transform}{Whether to log2 transform the data}

\item{distance_method}{Distance method for MDS}
}
\value{
ggplot2 object
}
\description{
MDS plot of gene expression data
}
