% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pca}
\alias{plot_gene_pca}
\title{PCA plot of gene expression data}
\usage{
plot_gene_pca(
  expr_matrix,
  sample_labels = NULL,
  title = "PCA of Gene Expression",
  log_transform = FALSE
)
}
\arguments{
\item{expr_matrix}{Gene expression matrix (genes as rows, samples as columns)}

\item{sample_labels}{Optional vector of sample labels}

\item{title}{Plot title}

\item{log_transform}{Whether to log2 transform the data}
}
\value{
ggplot2 object
}
\description{
PCA plot of gene expression data
}
\examples{
\dontrun{
plot_gene_pca(expr_matrix)
}
}
