% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_tsne}
\alias{plot_gene_tsne}
\title{t-SNE plot of gene expression data}
\usage{
plot_gene_tsne(
  expr_matrix,
  sample_labels = NULL,
  title = "t-SNE of Gene Expression",
  log_transform = FALSE,
  perplexity = 5
)
}
\arguments{
\item{expr_matrix}{Gene expression matrix (genes as rows, samples as columns)}

\item{sample_labels}{Optional vector of sample labels}

\item{title}{Plot title}

\item{log_transform}{Whether to log2 transform the data}

\item{perplexity}{Perplexity parameter for t-SNE}
}
\value{
ggplot2 object
}
\description{
t-SNE plot of gene expression data
}
